requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
attrs==25.1.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
curlify==2.2.1
execnet==2.1.1
Faker==37.0.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
playwright==1.50.0
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
pytest==8.3.4
requests==2.32.3
SQLAlchemy==2.0.39
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
#